Feature/enable dripper configuration (#7)
* Added dripper configuration object.

This commit abstracts the dripper configuration from constants to a config
object with a function to generate a config object with sane defaults.

* Added development deploy script.

This commit adds a script to build and deploy the cold brew applciation for
local development. This allows all development to be done on workstation and not
require making changes and compiling directly on the raspberry pi.

* Added base dripper tests with mock dripper.

This commit adds a mock dripper package and setups a base testing framework.
These tests will need to be revisited, but I at least wanted to commit what I
have.

* Added dripper configuration endpoint.

This commit adds a configuration endpoint to reinitialize the dripper with a new
configuration while the applicaiton is running.

* Added settings page.

This commit adds a settings page that allows for detailed configuration settings
to be set on the dripper.