Feature/add filesystem based database (#11)
* Added base filesystem database.

This commit adds the base wiring for a filesystem based database. This
approach was chosen due to sqlites C dependencies and the additional
complications for cross complilations. A full blown SQL database was
also considered, however, I didn't an external dependency that users
would have to instal.

* Refactored dripper config to settings.

This commit refactors dripper config to be dripper settings to avoid
confusion with the application configuration.

* Added settings to database.

This commit adds settings to the database. This allows users to persist
their dripper settings between restarts of the cold-brew application.

* Removed dep ensure from build.

This commit removes dep ensure from the build. The logic here is that
the build should fail if the dependencies are not vendored properly
opposed to the current situation where the dependencies will be
downloaded locally, built, but never committed.