Mspicer/get it working again (#15)
* Updated local dev frontend.

This commit udpates the local dev backend location.

* Fixed binary location for service.

This commit updates the binary location for the service.

* Fixed issue with hat driver.

This commit fixes an issue with the motor hat driver. It's a bit janky, but it's
definitely the most simple and reliable fix for the short term.

* Updated readme for common operations.

* Removed electron.

This feature was half baked and shouldn't have been merged.